+++ src/pages/[chapter]/[word].astro
@@ -2,8 +2,8 @@
 import type { GetStaticPathsResult } from "astro";
 import Audio from "../../components/Audio.astro";
 import Navigation from "../../components/Navigation.astro";
-import { dictionary } from "../../dictionary";
 import Layout from "../../layouts/Layout.astro";
+import { dictionary } from "../../lib/dictionary";
 
 export function getStaticPaths() {
 	return Object.values(dictionary.chapters).reduce<GetStaticPathsResult>(
