+++ src/pages/index.astro
@@ -42,7 +42,7 @@ import Layout from "../layouts/Layout.astro";
 
 <script>
 	import { animate, stagger } from "motion";
-	import { expoInOut } from "../util";
+	import { expoInOut } from "../lib/util";
 
 	document.addEventListener("astro:page-load", () => {
 		animate([
